Add integration tests for authorization (#130)
Added integration tests for the authorization functionality of the
Zenoh transport. The tests cover scenarios for both authorized and
unauthorized clients, ensuring that the authorization mechanism
works as expected. The tests also verify that the access control
configuration is correctly applied, and that clients can only
publish and subscribe to topics they are authorized to access.

Also changed CI jobs to no longer use nextest, as it is not
compatible with the test-serial crate, which is used for the
authorization integration tests.